Split generic behavior requires tests
