+++ src/typechecker/tests/generic_behaviors.rs
@@ -5,3 +5,4 @@ mod extends;
 mod generic_bounds;
 mod generic_function_diagnostics;
 mod impls_and_requires;
+mod requires;
